unwrangle gherkin statement matching inside implemetation code
