+++ features/PSE001_Standard-properties-and-property-sets-validation.feature
@@ -9,11 +9,13 @@ The rule verifies that each IfcPropertySet starting with Pset is defined correct
    Given An IfcPropertySet
    Given its Name attribute starts with Pset
 
+
     Scenario Outline: IfcPropertySet Name
      
       Given a file with Schema "<schema>"
+      Given its attribute Name
 
-      Then The IfcPropertySet Name attribute value must use predefined values according to the "<csv_table>" table
+      Then It must use predefined values according to the "<csv_table>" table
 
       Examples:
         | schema  | csv_table |
@@ -25,8 +27,10 @@ The rule verifies that each IfcPropertySet starting with Pset is defined correct
     Scenario Outline: Property Name
 
         Given a file with Schema "<schema>"
+        Given its attribute HasProperties
+        Given its attribute Name
 
-        Then Each associated IfcProperty must be named according to the property set definitions table "<csv_table>"
+        Then It must be named according to the property set definitions table "<csv_table>"
 
         Examples:
           | schema  | csv_table |
@@ -51,8 +55,9 @@ The rule verifies that each IfcPropertySet starting with Pset is defined correct
     Scenario Outline: Property Type
 
         Given a file with Schema "<schema>"
+        Given its attribute HasProperties
 
-        Then Each associated IfcProperty must be of type according to the property set definitions table "<csv_table>"
+        Then It must be of type according to the property set definitions table "<csv_table>"
 
         Examples:
           | schema  | csv_table |
@@ -60,15 +65,16 @@ The rule verifies that each IfcPropertySet starting with Pset is defined correct
           | IFC4    | IFC4_definitions.csv |
           | IFC4X3  | IFC4X3_definitions.csv |
 
-      
+
     Scenario Outline: Property Data Type
-      
-          Given a file with Schema "<schema>"
-  
-          Then Each associated IfcProperty value must be of data type according to the property set definitions table "<csv_table>"
-  
-          Examples:
-            | schema  | csv_table |
-            | IFC2X3  | IFC2x3_definitions.csv |
-            | IFC4    | IFC4_definitions.csv |
-            | IFC4X3  | IFC4X3_definitions.csv |
+    
+        Given a file with Schema "<schema>"
+        Given its attribute HasProperties
+
+        Then It must be of data type according to the property set definitions table "<csv_table>"
+
+        Examples:
+          | schema  | csv_table |
+          | IFC2X3  | IFC2x3_definitions.csv |
+          | IFC4    | IFC4_definitions.csv |
+          | IFC4X3  | IFC4X3_definitions.csv |
