+++ features/steps/steps.py
@@ -1,3 +1,3 @@
 from givens import attributes, entities, relationships, values
-from thens import alignment, attributes, existance, geometry, nesting, reference, relations, values
+from thens import alignment, attributes, existance, geometry, nesting, reference, relations, values, psets
 from steps import attributes
